feat: add auth router to core
