+++ central-oon-core-backend/package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "axios": "^1.7.7",
     "dotenv": "^16.4.5",
+    "express": "^4.21.2",
     "multer": "^1.4.5-lts.1",
     "winston": "^3.17.0"
   }
